+++ Services/Services.Contracts/SonarTask/CreatingSearchTaskDto.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// Идентификатор процесса поиска.
         /// </summary>
-        public int SonarProcessId { get; set; }
+        public long SonarProcessId { get; set; }
 
         /// <summary>
         /// Тема.
